locality fields (state, country) to Charfield, and LocalityField class
state field from json to charfield, custom AddressStateField

convert StateAddressField to Locality field that handles state and country

reorg manifest related serializers into package

WasteLineSerializer now child of TrakSerializer
