+++ apps/api/serializers/__init__.py
@@ -1,4 +1,3 @@
-# from .addresses import AddressSerializer, MailAddressField, SiteAddressField
 from .address import AddressSerializer
 from .handler import HandlerSerializer
 from .manifest import (ManifestSerializer, TransporterSerializer,
