+++ apps/api/serializers/address.py
@@ -1,16 +1,41 @@
 from rest_framework import serializers
+from rest_framework.exceptions import ValidationError
 
 from apps.trak.models import Address
+from lib.rcrainfo.lookups import get_country_name, get_state_name
 
 from .base import TrakSerializer
 
-# from lib.rcrainfo.lookups import get_country_name, get_state_name
+
+class LocalityField(serializers.Field):
+
+    def __init__(self, locality_type):
+        super().__init__()
+        self.locality_type = locality_type
+        self.locality_name = None
+
+    def to_internal_value(self, data):
+        try:
+            return data['code']
+        except KeyError:
+            raise ValidationError(f'state code is required for address {data}')
+
+    def to_representation(self, value) -> dict:
+        if self.locality_type == 'state':
+            self.locality_name = get_state_name(value)
+        elif self.locality_type == 'country':
+            self.locality_name = get_country_name(value)
+        representation = {"code": value,
+                          "name": self.locality_name}
+        return representation
 
 
 class AddressSerializer(TrakSerializer):
     streetNumber = serializers.CharField(
         source='street_number',
     )
+    state = LocalityField('state')
+    country = LocalityField('country')
 
     class Meta:
         model = Address
