+++ apps/api/serializers/manifest/__init__.py
@@ -1,0 +1,2 @@
+from .manifest import ManifestSerializer, WasteLineSerializer
+from .transporter import TransporterSerializer
